Compute correction pattern spherical aberrations
By ray tracing a point source to the camera plane, a correction pattern
based on the pathlength can be computed to counter spherical
aberrations. By changing the point source position, the amount of
defocus can be adjusted. This allows for yielding a pattern more
suitable for a real SLM, with a finite resolution.

Numpy and Scipy dependencies added.