Add 3D and 4D fit test
To model the distortion of the optical system itself, the x,y,kx,ky of
the camera images will be mapped to those of the SLM and Galvo Mirrors.
This test demonstrates how a 3D or 4D polynomial can be fit to a set op
in/output coordinates, by using random coefficients for the polynomial +
noise. For the actual distortions, not all polynomial basis vectors
might be relevant.